test/vti/test_serving/gae/requirements.txt
google-api-python-client
google-endpoints

pytz
stripe

